normalizado url, corregido un get, eliminado listas sin usar(algunas faltan buscar otras),creado listaEmpreaSprints, implementado en todas las listas el componenteListaDefinitivaN
