+++ Backend/app/Http/Controllers/Empresa/PlanificacionController.php
@@ -58,10 +58,10 @@ class PlanificacionController extends Controller
                 ->first();
 
             // Verificar si la planificación existe y si fue rechazada
-            if ($planificacion && $planificacion->publicada == 1 && $planificacion->aceptada == 0) {
+            if ($planificacion && $planificacion->aceptada === 0) {
                 // Si la planificación existe y fue rechazada, guarda sus datos
                 $data[] = [
-                    'idPlanificacion' => $planificacion->idPlanificacion,
+                    'id' => $planificacion->idPlanificacion,
                     'nombreEmpresa' => $empresa->nombreEmpresa,
                     'nombreLargo' => $empresa->nombreLargo,
                     'idEmpresa' => $planificacion->idEmpresa,
