+++ Frontend/src/pages/Docente/calificarSprint/listaSprints.jsx
@@ -12,7 +12,7 @@ const ListaSprints = () => {
     const navigate = useNavigate();
     const [sprints, setSprints] = useState([
     ])
-    const { idEmpresa } = useParams();
+    const { idEmpresa, idGrupo } = useParams();
     const [loading, setLoading] = useState(true);  
     const [error, setError] = useState({
         error: false,
@@ -43,7 +43,7 @@ const ListaSprints = () => {
     },[])
     const clickBoton = (sprint) => {
         console.log("click", sprint)
-        navigate(`/${idEmpresa}/calificarSprints/sprint/${sprint}`); 
+        navigate(`homeGrupo/${idGrupo}/listaEmpresaCalificarSprints/${idEmpresa}/sprint/${sprint}`); 
     }
     if (loading) {
         return <Loading></Loading>
@@ -56,7 +56,7 @@ const ListaSprints = () => {
             titulo={'SELECCIONE UN SPRINT PARA CALIFICAR'}
             ocultarAtras={false}
             confirmarAtras={false}
-            dirBack={`/`}
+            dirBack={`homeGrupo/${idGrupo}/listaEmpresaCalificarSprints`}
         >
         <DivLista>
             {sprints!==null?
