+++ Frontend/src/pages/Estudiante/modificarTarea/listaTareas.jsx
@@ -1,11 +1,11 @@
 import { useParams } from 'react-router-dom';
 import VisualizarSprintEst from '../visualizarSprintEstudiante/visualizarSprint';
 function ListaTareas() {
-    const { idSprint } = useParams(); 
+    const { idSprint, idEstudiante, idGrupo } = useParams(); 
   return (
     <VisualizarSprintEst 
       titulo={'SELECCIONE UNA TAREA PARA MODIFICAR'} 
-      navigateLink={`/homeEstudiante/homeGrupoEstudiante/sprintE/${idSprint}/semana/editarTarea/`}
+      navigateLink={`/${idEstudiante}/homeGrupoE/${idGrupo}/sprintE/${idSprint}/semana/editarTarea/`}
     ></VisualizarSprintEst>
   );
 }
